Adds Page porting button. Renames ManagePages to PageLibrary
