+++ apps/modernization-ui/src/apps/page-builder/components/QuickConditionLookup/QuickConditionLookup.spec.tsx
@@ -2,7 +2,7 @@ import { render, screen, act, fireEvent, waitFor } from '@testing-library/react'
 import { QuickConditionLookup } from './QuickConditionLookup';
 import { PagesContext } from 'apps/page-builder/context/PagesContext';
 import { Direction } from 'sorting';
-import { Column } from 'apps/page-builder/pages/ManagePages/ManagePagesTable';
+import { Column } from 'apps/page-builder/pages/PageLibrary/ManagePagesTable';
 import { BrowserRouter } from 'react-router-dom';
 import { ConditionControllerService } from 'apps/page-builder/generated/services/ConditionControllerService';
 import { ModalRef } from '@trussworks/react-uswds';
