+++ apps/modernization-ui/src/apps/page-builder/pages/PageLibrary/ManagePagesTable.scss
