+++ apps/modernization-ui/src/routes/AppRoutes.tsx
@@ -3,7 +3,7 @@ import { AdvancedSearch } from '../pages/advancedSearch/AdvancedSearch';
 import { Login } from '../pages/login/Login';
 import { PatientProfile } from '../pages/patient/profile';
 import { AddPatient } from 'pages/addPatient/AddPatient';
-import { ManagePages } from 'apps/page-builder/pages/ManagePages/ManagePages';
+import { PageLibrary } from 'apps/page-builder/pages/PageLibrary/PageLibrary';
 import { AddNewPage } from 'apps/page-builder/pages/AddNewPage/AddNewPage';
 import { ReactNode, useContext, useEffect, useState } from 'react';
 import { UserContext } from 'providers/UserContext';
@@ -75,7 +75,7 @@ export const AppRoutes = () => {
                             {config.features.pageBuilder.enabled ? (
                                 <Route path="/page-builder" element={<PageBuilderContextProvider />}>
                                     <Route path="manage">
-                                        <Route path="pages" element={<ManagePages />} />
+                                        <Route path="pages" element={<PageLibrary />} />
                                         <Route path="valueset-library" element={<ValuesetLibrary />} />
                                         <Route path="question-library" element={<QuestionLibrary />} />
                                         <Route path="business-rules-library" element={<BusinessRulesLibrary />} />
